Improve global_mem_cache_size and local_mem_size
This code sets global_mem_cache_size equal to the size of the last
level cache (typically L3 on current CPUs) and local_mem_size equal to
the size of the last private cache of a core (typically L2 on current
CPUs) divided by the number of hardware threads for that cache. Move the
specification of local_mem_size from common.c to pocl_topology.c.